InputCalcItem.storyboardの追加 ExtensionsフォルダとFloatが金額表示を出力できる仕組みを作りました。 金額表示についてはテストも書きます。

+++ SumCalculator/Model/CalcItem.swift
@@ -10,10 +10,10 @@ import Foundation
 class CalcItem {
     let id: String = ""
     var name: String = ""
-    var unitPrice: Int = 0
-    var quantity: Int = 0
+    var unitPrice: Float = 0
+    var quantity: Float = 0
     
-    var subtotal: Int {
+    var subtotal: Float {
         get {
             return unitPrice * quantity
         }
